Move sortedAttributeOptions behaviors in a custom hook
