+++ src/Akeneo/Pim/Structure/Bundle/Resources/public/js/attribute-option/components/List.tsx
@@ -1,12 +1,14 @@
 import React, {useEffect, useState} from 'react';
+import {useSelector} from 'react-redux';
+
+import {useTranslate} from '@akeneo-pim-community/legacy-bridge';
 import {AttributeOption} from '../model';
+import {useAttributeContext} from '../contexts';
+import {AttributeOptionsState} from '../store/store';
+import {useSortedAttributeOptions} from '../hooks/useSortedAttributeOptions';
 import ToggleButton from './ToggleButton';
 import ListItem, {DragItem} from './ListItem';
-import {useTranslate} from '@akeneo-pim-community/legacy-bridge';
 import NewOptionPlaceholder from './NewOptionPlaceholder';
-import {useAttributeContext} from '../contexts';
-import {useSelector} from 'react-redux';
-import {AttributeOptionsState} from '../store/store';
 
 interface ListProps {
     selectAttributeOption: (selectedOptionId: number | null) => void;
@@ -21,8 +23,12 @@ const List = ({selectAttributeOption, selectedOptionId, isNewOptionFormDisplayed
     const attributeOptions = useSelector((state: AttributeOptionsState) => state.attributeOptions);
     const translate = useTranslate();
     const attributeContext = useAttributeContext();
+    const {
+        sortedAttributeOptions,
+        moveAttributeOption,
+        validateMoveAttributeOption
+    } = useSortedAttributeOptions(attributeOptions, attributeContext.autoSortOptions, manuallySortAttributeOptions);
     const [showNewOptionPlaceholder, setShowNewOptionPlaceholder] = useState<boolean>(isNewOptionFormDisplayed);
-    const [sortedAttributeOptions, setSortedAttributeOptions] = useState<AttributeOption[] | null>(attributeOptions);
     const [dragItem, setDragItem] = useState<DragItem | null>(null);
 
     useEffect(() => {
@@ -31,18 +37,6 @@ const List = ({selectAttributeOption, selectedOptionId, isNewOptionFormDisplayed
         }
     }, [selectedOptionId]);
 
-    useEffect(() => {
-        if (attributeOptions !== null) {
-            let sortedOptions = [...attributeOptions];
-            if (attributeContext.autoSortOptions) {
-                // /!\ sort() does not return another reference, it sorts directly on the original variable.
-                sortedOptions.sort((option1: AttributeOption, option2: AttributeOption) => {
-                    return option1.code.localeCompare(option2.code, undefined, {sensitivity: 'base'});
-                });
-            }
-            setSortedAttributeOptions(sortedOptions);
-        }
-    }, [attributeOptions, attributeContext.autoSortOptions]);
 
     const onSelectItem = (optionId: number) => {
         setShowNewOptionPlaceholder(false);
@@ -61,26 +55,6 @@ const List = ({selectAttributeOption, selectedOptionId, isNewOptionFormDisplayed
         setShowNewOptionPlaceholder(false);
         if (attributeOptions !== null && attributeOptions.length > 0) {
             selectAttributeOption(attributeOptions[0].id);
-        }
-    };
-
-    const moveAttributeOption = (sourceOptionCode: string, targetOptionCode: string) => {
-        if (sortedAttributeOptions !== null && sourceOptionCode !== targetOptionCode) {
-            const sourceIndex = sortedAttributeOptions.findIndex((attributeOption: AttributeOption) => attributeOption.code === sourceOptionCode);
-            const targetIndex = sortedAttributeOptions.findIndex((attributeOption: AttributeOption) => attributeOption.code === targetOptionCode);
-            const sourceOption = sortedAttributeOptions[sourceIndex];
-
-            let newSortedAttributeOptions = [...sortedAttributeOptions];
-            newSortedAttributeOptions.splice(sourceIndex, 1);
-            newSortedAttributeOptions.splice(targetIndex, 0, sourceOption);
-
-            setSortedAttributeOptions(newSortedAttributeOptions);
-        }
-    };
-
-    const validateMoveAttributeOption = () => {
-        if (sortedAttributeOptions !== null && JSON.stringify(sortedAttributeOptions) != JSON.stringify(attributeOptions)) {
-            manuallySortAttributeOptions(sortedAttributeOptions);
         }
     };
