+++ src/Akeneo/Pim/Structure/Bundle/Resources/public/js/attribute-option/hooks/index.ts
@@ -6,6 +6,7 @@ import {useDeleteAttributeOption} from './useDeleteAttributeOption';
 import {useManualSortAttributeOptions} from './useManualSortAttributeOptions';
 
 export * from './useEditingOptionContextState';
+export * from './useSortedAttributeOptions';
 
 export {
     useAttributeOptions,
